+++ Changelog.md
@@ -121,3 +121,11 @@ Contains information about the versions of Termsequel
         | :-:   | :-: | :-: |
         | Windows, does not have the implemention of [getopt](https://www.man7.org/linux/man-pages/man3/getopt.3.html), so, we could not use for example: `$ termsequel -h` to see the help ... | [ISSUE](https://github.com/sgtcortez/Termsequel/issues/88) | [PR](https://github.com/sgtcortez/Termsequel/pull/90)  |
         | Max level is now a command line argument, instead of a filter condition | [ISSUE](https://github.com/sgtcortez/Termsequel/issues/80) | [PR](https://github.com/sgtcortez/Termsequel/pull/92) |
+
+    - Features
+
+        There is a list of features from this release.
+
+        | Feature | Issue | Pull Request |
+        | :-:   | :-: | :-: | :-: |
+        | Add GROUP column for linux systems | [ISSUE](https://github.com/sgtcortez/Termsequel/issues/75) | [PR](https://github.com/sgtcortez/Termsequel/pull/97) |